complemento de 2 foi de 8bits para 16bits

+++ src/encoding.h
@@ -8,4 +8,4 @@ int binaryToUnsignedInt(const char *serializedBinary);
 
 void extendSignal(char *baseSerializedBinary, char *buffer, int size);
 
-int8_t complementOfTwoToInt(const char *serializedBinary);
+int16_t complementOfTwoToInt(const char *serializedBinary);
